Initial implementation of browser annotation flow
Allows reviewers to select text, add a comment, and push it to the global state.  Can handle overlapping highlights. There may be edge cases (literally at the edges especially). The implementation also also screws up the selection, so it’s not possible to copy text. I’ll add a subsequent bug for this.

Fixes #16